+++ tests/test_un_cli.py
@@ -1,5 +1,6 @@
-from .conftest import pytest
 from udc import UnCli
+
+from .conftest import BENCH_URI, PKG_URI, pytest
 
 
 @pytest.fixture
@@ -14,7 +15,12 @@ def test_un_cli(config: UnCli):
 
 
 def test_un_cli_commands(config: UnCli):
-    cf_list = config.command('list')
+    cf_list = config.cmd_opts("list")
     assert cf_list
-    assert 'name' in cf_list
-    assert cf_list['name'] == 'list'
+    assert "name" in cf_list
+    assert cf_list["name"] == "list"
+
+
+def test_un_get_resource(config: UnCli):
+    assert config.get_resource(PKG_URI)
+    assert config.get_resource(BENCH_URI)