+++ tools/benchling/__init__.py
@@ -1,0 +1,5 @@
+from .entry import BenchlingEntry, BenchlingEntryList  # NOQA F401
+from .resource import BenchlingResource, RESOURCE_MAP  # NOQA F401
+from .root import BenchlingRoot, BenchlingById  # NOQA F401
+from .schema import BenchlingSchema, BenchlingSchemaList  # NOQA F401
+from .sequence import BenchlingSequence, BenchlingSequenceList  # NOQA F401