+++ tools/quilt/__init__.py
@@ -1,0 +1,1 @@
+from .resource import QuiltResource  # NOQA F401