+++ udc/config.py
@@ -6,8 +6,7 @@ from datetime import datetime
 from pathlib import Path
 
 import yaml
-
-from .api.id import QuiltID
+from quiltplus import QuiltID
 
 __version__ = importlib.metadata.version(__package__)